+++ src/parallelalgo.h
@@ -14,8 +14,8 @@ double parallelSquaredNorm(const VectorXd &epsilon);
 
 // Inner loop helpers
 void parallelCastDouble(const VectorXf &in, VectorXd &out);
-void parallelUpdateYTilde(VectorXd &y_tilde, const VectorXd &epsilon, const VectorXf &Cx, double beta);
-void parallelUpdateEpsilon(VectorXd &epsilon, const VectorXd &y_tilde, const VectorXf &Cx, double beta);
-double parallelDotProduct(const VectorXf &Cx, const VectorXd &y_tilde);
+void parallelUpdateYTilde(VectorXd &y_tilde, const VectorXd &epsilon, const VectorXd &Cx, double beta);
+void parallelUpdateEpsilon(VectorXd &epsilon, const VectorXd &y_tilde, const VectorXd &Cx, double beta);
+double parallelDotProduct(const VectorXd &Cx, const VectorXd &y_tilde);
 
 #endif // PRALLELALGO_H
